// Minhas Api Para o vitor/Typescritp-Scan-Funcionarios/src/types/funcionarioTypes.ts
 export type FuncionarioType = {
    nome: string;
    email: string;
    telefone: string;
    cpf: string;
    cep: string;
    cidade: string;
    estado: string;
    cargo: string;
    salario: number;
    curriculo: string;
    ensino_medio: boolean;
    ensino_superior: boolean;
    pos_graduacao: boolean;
    mestrado: boolean;
    nota_teste: number;
    nota_entrevista: number;
};